Set minimum deployment target to iOS 12
The default minimum iOS target used to be iOS 11, and while PojavLauncher would *run*, the experience would be terrible at beast. Microsoft authentication would crash the app, as the iOS 11 methods were deprecated and replace in iOS 12, and OpenJDK would be missing features as 1) Procursus isn't supported on iOS 11 and 2) the JDK targets iOS 12 and higher. This commit forces the scripts to build for iOS 12 at the minimum so iOS 11 doesn't give off errors during compile time and when installing.