2.1.1 - Fix issue with sidebar on iOS 12 and 13
Swapping between OneOver and OneBeside confuses iOS 12/13 and caused issues with opening the sidebar after it has been closed. With this fix, iOS 12 and 13 devices receive a hidden sidebar with no option to change this behavior, whereas iOS 14 and up keep the same behavior. When this is ported to the 2.2 code, there will be an additional option for iOS 14 and up users to swap between a hidden and docked sidebar when in landscape mode.