An actual workaround for the UI initialization bug without forcing the iPad UI
Basically consists of pushing the LauncherNewsViewController from LauncherNavigationController when the app first starts up, and setting the navigation bar items from the LNVC class. This commit also moves the account button to the right of the bar, removes the profile name outside of the list view, and does not appear on the Settings view. Removes the hidden sidebar option in favor of always hidden. Moves PID to the bottom of the General section footer. Tested on my iPad Pro (12.9-inch, 2nd generation), iPhone 14 Pro Max, iPhone 6s Plus, and iPhone SE (1st generation) over a grueling week of staring at code and banging my head against a wall. I'm going to bed now.